requirements.txt
python-telegram-bot==22.5
flask==3.0.0
requests==2.31.0
gunicorn==21.2.0